requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
pydantic==1.10.13
httpx==0.25.2
schedule==1.2.0
yfinance==0.2.35
Werkzeug==3.0.1
PyJWT==2.8.0
email-validator==2.1.0